feat: Add schema extraction functionality with DDL and Markdown output
- Introduced a new command `schema` to extract database schema from Access files.
- Implemented schema reading logic in the internal `mdb` package.
- Added support for generating DDL and Markdown representations of the schema.
- Included tests for schema reading and Jet type to SQL mapping.
- Removed obsolete cross-validation documentation.

+++ internal/vba/decompress_test.go
@@ -23,7 +23,7 @@ func (h *captureHandler) Handle(_ context.Context, r slog.Record) error {
 func buildContainerChunk(compressed bool, payload []byte) []byte {
 	chunkSize := len(payload) + 2
 
-	base := uint16(0x3000 | uint16(chunkSize-3))
+	base := 0x3000 | uint16(chunkSize-3)
 	if compressed {
 		base |= 0x8000
 	}
